Kept only ci_sanity tests for heat tests
